+++ cmd/soroban-rpc/internal/methods/get_events_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"math"
 	"path"
 	"strconv"
 	"strings"
@@ -648,14 +649,14 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               "CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABSC4",
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{value},
 				ValueXDR:                 value,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(i).HexString(),
 			})
 		}
-		assert.Equal(t, GetEventsResponse{expected, 1}, results)
+		cursor := db.Cursor{Ledger: 1, Tx: math.MaxUint32, Event: math.MaxUint32 - 1}.String()
+		assert.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 	})
 
 	t.Run("filtering by contract id", func(t *testing.T) {
@@ -794,14 +795,15 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               "CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABSC4",
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{counterXdr, value},
 				ValueXDR:                 value,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(4).HexString(),
 			},
 		}
-		assert.Equal(t, GetEventsResponse{expected, 1}, results)
+		cursor := db.Cursor{Ledger: 1, Tx: math.MaxUint32, Event: math.MaxUint32 - 1}.String()
+
+		assert.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 
 		results, err = handler.getEvents(ctx, GetEventsRequest{
 			StartLedger: 1,
@@ -835,7 +837,7 @@ func TestGetEvents(t *testing.T) {
 
 		expected[0].ValueJSON = valueJs
 		expected[0].TopicJSON = topicsJs
-		require.Equal(t, GetEventsResponse{expected, 1}, results)
+		require.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 	})
 
 	t.Run("filtering by both contract id and topic", func(t *testing.T) {
@@ -939,14 +941,15 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               strkey.MustEncode(strkey.VersionByteContract, contractID[:]),
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{counterXdr, value},
 				ValueXDR:                 value,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(3).HexString(),
 			},
 		}
-		assert.Equal(t, GetEventsResponse{expected, 1}, results)
+		cursor := db.Cursor{Ledger: 1, Tx: math.MaxUint32, Event: math.MaxUint32 - 1}.String()
+
+		assert.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 	})
 
 	t.Run("filtering by event type", func(t *testing.T) {
@@ -1014,14 +1017,15 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               strkey.MustEncode(strkey.VersionByteContract, contractID[:]),
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{counterXdr},
 				ValueXDR:                 counterXdr,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(0).HexString(),
 			},
 		}
-		assert.Equal(t, GetEventsResponse{expected, 1}, results)
+		cursor := db.Cursor{Ledger: 1, Tx: math.MaxUint32, Event: math.MaxUint32 - 1}.String()
+
+		assert.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 	})
 
 	t.Run("with limit", func(t *testing.T) {
@@ -1085,14 +1089,15 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               "CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABSC4",
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{value},
 				ValueXDR:                 value,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(i).HexString(),
 			})
 		}
-		assert.Equal(t, GetEventsResponse{expected, 1}, results)
+		cursor := expected[len(expected)-1].ID
+
+		assert.Equal(t, GetEventsResponse{expected, 1, cursor}, results)
 	})
 
 	t.Run("with cursor", func(t *testing.T) {
@@ -1185,14 +1190,14 @@ func TestGetEvents(t *testing.T) {
 				LedgerClosedAt:           now.Format(time.RFC3339),
 				ContractID:               strkey.MustEncode(strkey.VersionByteContract, contractID[:]),
 				ID:                       id,
-				PagingToken:              id,
 				TopicXDR:                 []string{counterXdr},
 				ValueXDR:                 expectedXdr,
 				InSuccessfulContractCall: true,
 				TransactionHash:          ledgerCloseMeta.TransactionHash(i).HexString(),
 			})
 		}
-		assert.Equal(t, GetEventsResponse{expected, 5}, results)
+		cursor := expected[len(expected)-1].ID
+		assert.Equal(t, GetEventsResponse{expected, 5, cursor}, results)
 
 		results, err = handler.getEvents(context.TODO(), GetEventsRequest{
 			Pagination: &PaginationOptions{
@@ -1201,7 +1206,14 @@ func TestGetEvents(t *testing.T) {
 			},
 		})
 		require.NoError(t, err)
-		assert.Equal(t, GetEventsResponse{[]EventInfo{}, 5}, results)
+
+		latestLedger := 5
+		endLedger := min(5+LedgerScanLimit, latestLedger+1)
+
+		// Note: endLedger is always exclusive when fetching events
+		// so search window is always max Cursor value with endLedger - 1
+		cursor = db.Cursor{Ledger: uint32(endLedger - 1), Tx: math.MaxUint32, Event: math.MaxUint32 - 1}.String()
+		assert.Equal(t, GetEventsResponse{[]EventInfo{}, 5, cursor}, results)
 	})
 }